BUG: remove an unneeded layer of pointers
ptrf is actually an index into the offsets array, not a pointer.  (See
CASE_ERODE_POINT2.)  So the element type size must not be (implicitly)
added to it here, when incrementing the value.  The element type size
will be (implicitly) added later, when accessing the offsets array with
square brackets.